+++ test/PlannerSpec.hs
@@ -28,6 +28,7 @@ testPlanner :: TravelPreferences -> Camino -> Test
 testPlanner preferences camino = TestList [
   TestLabel "Hours Simple" testHoursSimple,
   TestLabel "Travel Simple" testTravelSimple,
+  TestLabel "Accommodation Map Simple" testAccommodationMapSimple,
   TestLabel "Accommodation Simple" testAccommodationSimple,
   TestLabel "Penance Simple" testPenanceSimple,
   TestLabel "Plan Camino" (testPlanCamino preferences camino)
@@ -43,7 +44,8 @@ preferences1 = TravelPreferences {
     preferenceStop = Penance 0.5,
     preferenceAccommodation = M.fromList [
         (MunicipalAlbergue, (Penance 1.5)),
-        (PrivateAlbergue, (Penance 0.9))
+        (PrivateAlbergue, (Penance 0.9)),
+        (Hotel, (Penance 0.5))
       ],
     preferenceStopServices = M.fromList [
       (Restaurant, (Penance 0.2)),
@@ -63,8 +65,8 @@ preferences2 = TravelPreferences {
     preferenceTime = PreferenceRange Nothing 6.0 0.0 8.0 Nothing (Just 10.0),
     preferenceStop = Penance 0.5,
     preferenceAccommodation = M.fromList [
-        (MunicipalAlbergue, (Penance 0.0)),
-        (PrivateAlbergue, (Penance 1.0))
+        (MunicipalAlbergue, (Penance 4.0)),
+        (PrivateAlbergue, (Penance 3.0))
       ],
     preferenceStopServices = M.fromList [
       (Restaurant, (Penance 0.2)),
@@ -129,8 +131,6 @@ location4 = Location {
     ]  
   }
   
-services0 = S.fromList [ Restaurant, Groceries ]
-
 
 legs0 = [
   Leg { legType = Road, legFrom = location1, legTo = location2, legDistance = 2.0, legTime = Nothing, legAscent = 100, legDescent = 50, legPenance = Nothing, legNotes = Nothing }
@@ -175,6 +175,8 @@ cpreferences1 = CaminoPreferences {
   preferenceStops = S.empty,
   preferenceExcluded = S.empty
 }
+
+accommodationMap1 = buildAccommodationMap preferences1 camino1 location1 location3 (const True)
   
 testHoursSimple = TestList [testHoursSimple1, testHoursSimple2, testHoursSimple3, testHoursSimple4, testHoursSimple5, testHoursSimple6]
 
@@ -199,45 +201,57 @@ testTravelSimple2 = TestCase (assertFloatEqual "Travel Simple 2" 10.0 (travel pr
 
 testTravelSimple3 = TestCase (assertFloatEqual "Travel Simple 3" 2.0 (travel preferences1 legs0) 0.001)
 
-testAccommodationSimple = TestList [ testAccommodationSimple1, testAccommodationSimple2, testAccommodationSimple3 ]
+testAccommodationMapSimple = TestList [ testAccommodationMapSimple1 ]
+
+testAccommodationMapSimple1 = TestCase (do
+    assertEqual "Accommodation Map Simple 1 1" Reject (accommodationChoicePenance $ accommodationMap1 M.! location1)
+  )
+
+testAccommodationSimple = TestList [ testAccommodationSimple1, testAccommodationSimple2, testAccommodationSimple3, testAccommodationSimple4 ]
 
 testAccommodationSimple1 = let
-    (accom, serv, pen) = accommodation preferences1 camino1 legs0 services0 False
+    accom = accommodation preferences1 location2
   in
     TestCase (do
-      assertEqual "Accommodation Simple 1 1" True (isJust accom)
-      assertEqual "Accommodation Simple 1 2" PrivateAlbergue (accommodationType $ fromJust accom)
-      assertEqual "Accommodation Simple 1 3" services0 serv
-      assertPenanceEqual "Accommodation Simple 1 4" (Penance 1.4) pen 0.001
+      assertEqual "Accommodation Simple 1 1" MunicipalAlbergue (accommodationType $ accommodationChoice accom)
+      assertEqual "Accommodation Simple 1 2" S.empty (accommodationChoiceServices accom)
+      assertPenanceEqual "Accommodation Simple 1 3" (Penance 1.5) (accommodationChoicePenance accom) 0.001
     )
 testAccommodationSimple2 = let
-    (accom, serv, pen) = accommodation preferences1 camino1 legs1 services0 False
+    accom = accommodation preferences1 location1
   in
     TestCase (do
-      assertEqual "Accommodation Simple 2 1" True (isNothing accom)
-      assertEqual "Accommodation Simple 2 2" services0 serv
-      assertPenanceEqual "Accommodation Simple 2 3" Reject pen 0.001
+      assertEqual "Accommodation Simple 2 1" Camping (accommodationType $ accommodationChoice accom)
+      assertEqual "Accommodation Simple 2 2" S.empty (accommodationChoiceServices accom)
+      assertPenanceEqual "Accommodation Simple 2 3" Reject (accommodationChoicePenance accom) 0.001
     )
 testAccommodationSimple3 = let
-    (accom, serv, pen) = accommodation preferences2 camino1 legs0 services0 False
+    accom = accommodation preferences1 location3
   in
     TestCase (do
-      assertEqual "Accommodation Simple 3 1" True (isJust accom)
-      assertEqual "Accommodation Simple 3 2" MunicipalAlbergue (accommodationType $ fromJust accom)
-      assertEqual "Accommodation Simple 3 3" services0 serv
-      assertPenanceEqual "Accommodation Simple 3 4" (Penance 0.5) pen 0.001
+      assertEqual "Accommodation Simple 3 1" Hotel (accommodationType $ accommodationChoice accom)
+      assertEqual "Accommodation Simple 3 2" (S.singleton Restaurant) (accommodationChoiceServices accom)
+      assertPenanceEqual "Accommodation Simple 3 3" (Penance 0.5) (accommodationChoicePenance accom) 0.001
+    )
+testAccommodationSimple4 = let
+    accom = accommodation preferences2 location3
+  in
+    TestCase (do
+      assertEqual "Accommodation Simple 4 1" Hotel (accommodationType $ accommodationChoice accom)
+      assertEqual "Accommodation Simple 4 2" (S.singleton Restaurant) (accommodationChoiceServices accom)
+      assertPenanceEqual "Accommodation Simple 4 3" Reject (accommodationChoicePenance accom) 0.001
     )
 
 
 testPenanceSimple = TestList [ testPenanceSimple1, testPenanceSimple2, testPenanceSimple3, testPenanceSimple4 ]
 
-testPenanceSimple1 = TestCase (assertPenanceEqual "Penance Simple 1" (Penance 4.3) (metricsPenance $ penance preferences1 cpreferences1 legs0) 0.1)
+testPenanceSimple1 = TestCase (assertPenanceEqual "Penance Simple 1" (Penance 2.9) (metricsPenance $ penance preferences1 cpreferences1 accommodationMap1 legs0) 0.1)
 
-testPenanceSimple2 = TestCase (assertPenanceEqual "Penance Simple 2" (Penance 15.6) (metricsPenance $ penance preferences1 cpreferences1 legs1) 0.1)
+testPenanceSimple2 = TestCase (assertPenanceEqual "Penance Simple 2" (Penance 5.4) (metricsPenance $ penance preferences1 cpreferences1 accommodationMap1 legs1) 0.1)
 
-testPenanceSimple3 = TestCase (assertPenanceEqual "Penance Simple 3" (Penance 2.3) (metricsPenance $ penance preferences2 cpreferences1 legs0) 0.1)
+testPenanceSimple3 = TestCase (assertPenanceEqual "Penance Simple 3" (Penance 1.8) (metricsPenance $ penance preferences2 cpreferences1 accommodationMap1 legs0) 0.1)
 
-testPenanceSimple4 = TestCase (assertPenanceEqual "Penance Simple 4" (Penance 12.4) (metricsPenance $ penance preferences2 cpreferences1 legs1) 0.1)
+testPenanceSimple4 = TestCase (assertPenanceEqual "Penance Simple 4" (Penance 2.23) (metricsPenance $ penance preferences2 cpreferences1 accommodationMap1 legs1) 0.1)
 
 testPlanCamino preferences camino = TestList [ testPlanCamino1 preferences camino]
 
@@ -255,7 +269,7 @@ testPlanCamino1 preferences camino =
   in
     TestCase (do
       assertBool "Plan Camino 1 1" (isRight mroute)
-      let route = fromRight (error "Bad route") mroute
+      let route = fst $ fromRight (error "Bad route") mroute
       assertEqual "Plan Camino 1 2" (preferenceStart cpreferences) (start route)
       assertEqual "Plan Camino 1 3" (preferenceFinish cpreferences) (finish route)
       assertEqual "Plan Camino 1 4" 2 (length $ path route)
@@ -264,10 +278,10 @@ testPlanCamino1 preferences camino =
       assertEqual "Plan Camino 1 6" "P-P1" (identifier $ start day1)
       assertEqual "Plan Camino 1 7" "P-P7" (identifier $ finish day1)
       assertEqual "Plan Camino 1 8" 4 (length $ path day1)
-      assertPenanceEqual "Plan Camino 1 9" (Penance 26.5) (metricsPenance $ score day1) 0.1
+      assertPenanceEqual "Plan Camino 1 9" (Penance 24.0) (metricsPenance $ score day1) 0.1
       let day2 = path route !! 1
       assertEqual "Plan Camino 1 10" "P-P7" (identifier $ start day2)
       assertEqual "Plan Camino 1 11" "P-P12" (identifier $ finish day2)
       assertEqual "Plan Camino 1 12" 5 (length $ path day2)
-      assertPenanceEqual "Plan Camino 1 13" (Penance 23.5) (metricsPenance $ score day2) 0.1
+      assertPenanceEqual "Plan Camino 1 13" (Penance 23.0) (metricsPenance $ score day2) 0.1
     )